+++ lib/lighthouseProvider/backend/BLELighthouseBackend.dart
@@ -1,0 +1,8 @@
+import '../deviceProviders/BLEDeviceProvider.dart';
+import 'LighthouseBackend.dart';
+
+///
+/// A backend that will use [BLEDeviceProvider]s to provide [BLEDevice]s
+///
+abstract class BLELighthouseBackend
+    extends LighthouseBackend<BLEDeviceProvider> {}